Move Named & NamedMember into their own module.
