+++ src/Control/Effect/State/Named.hs
@@ -9,6 +9,7 @@ module Control.Effect.State.Named
 ) where
 
 import Control.Algebra
+import Control.Effect.Sum.Named
 import Control.Effect.State (State(..))
 
 -- | Get the current state value.
