+++ src/Control/Effect/Sum.hs
@@ -3,15 +3,11 @@ module Control.Effect.Sum
 ( (:+:)(..)
 , Member(..)
 , send
-, Named(..)
-, NamedMember(..)
-, sendNamed
 ) where
 
 import Control.Algebra.Class
 import Control.Effect.Class
 import GHC.Generics (Generic1)
-import GHC.TypeLits
 
 data (f :+: g) (m :: * -> *) k
   = L (f m k)
@@ -47,24 +43,3 @@ instance {-# OVERLAPPABLE #-} Member sub sup => Member sub (sub' :+: sup) where
 send :: (Member effect sig, Algebra sig m) => effect m a -> m a
 send = alg . inj
 {-# INLINE send #-}
-
-
-newtype Named (name :: Symbol) (eff :: (* -> *) -> (* -> *)) m k = Named { getNamed :: eff m k }
-
-class NamedMember (name :: Symbol) sub sup | name sup -> sub where
-  injNamed :: Named name sub m k -> sup m k
-
-instance NamedMember name sub (Named name sub) where
-  injNamed = id
-
-instance {-# OVERLAPPABLE #-} NamedMember name sub (Named name sub :+: sup) where
-  injNamed = L
-
-instance {-# OVERLAPPABLE #-} NamedMember name sub sup => NamedMember name sub (sub' :+: sup) where
-  injNamed = R . injNamed
-
-
--- | Construct a request for a named effect to be interpreted by some handler later on.
-sendNamed :: forall name effect sig m a . (NamedMember name effect sig, Algebra sig m) => effect m a -> m a
-sendNamed = alg . injNamed . Named @name
-{-# INLINE sendNamed #-}
